+++ src/java/org/ncbo/stanford/view/rest/restlet/subscriptions/SubscritpionsRestlet.java
@@ -9,6 +9,7 @@ import org.ncbo.stanford.service.subscriptions.SubscriptionsService;
 
 import org.ncbo.stanford.util.MessageUtils;
 import org.ncbo.stanford.util.RequestUtils;
+import org.ncbo.stanford.exception.NotificationNotFoundException;
 
 import org.ncbo.stanford.util.helper.BeanHelper;
 
@@ -20,6 +21,11 @@ import org.ncbo.stanford.domain.generated.NcboLNotificationType;
 import org.restlet.data.Request;
 import org.restlet.data.Response;
 import org.restlet.data.Status;
+
+/**
+ * @author g.prakash
+ * 
+ */
 
 public class SubscritpionsRestlet extends AbstractBaseRestlet {
 
@@ -109,14 +115,26 @@ public class SubscritpionsRestlet extends AbstractBaseRestlet {
 	private void createSubscriptions(Request request, Response response) {
 		SubscriptionsBean subscriptionsBean = BeanHelper
 				.populateSubscriptionsBeanFromRequest(request);
-		// create the subscriptions
+		// List for OntologyIds
+		List<String> ontologyId = subscriptionsBean.getOntologyIds();
 		try {
+			if (subscriptionsBean.getUserId() == null || ontologyId.isEmpty()
+					|| subscriptionsBean.getNotificationType() == null) {
+				throw new NotificationNotFoundException(
+						NotificationNotFoundException.DEFAULT_MESSAGE);
+			}
+			// create the subscriptions
 			subscriptionsService.createSubscriptions(subscriptionsBean);
+		} catch (NotificationNotFoundException notificationNotFoundException) {
+			response.setStatus(Status.CLIENT_ERROR_NOT_FOUND,
+					notificationNotFoundException.getMessage());
 		} catch (Exception e) {
 			response.setStatus(Status.SERVER_ERROR_INTERNAL, e.getMessage());
 			e.printStackTrace();
 
-		} finally {
+		}
+
+		finally {
 			// generate response XML
 			xmlSerializationService.generateXMLResponse(request, response,
 					subscriptionsBean);
@@ -132,9 +150,19 @@ public class SubscritpionsRestlet extends AbstractBaseRestlet {
 	private void deleteSubscriptions(Request request, Response response) {
 		SubscriptionsBean subscriptionsBean = BeanHelper
 				.populateSubscriptionsBeanFromRequest(request);
-		// create the subscriptions
+		// List For OntologyIds
+		List<String> ontologyId = subscriptionsBean.getOntologyIds();
+
 		try {
+			if (subscriptionsBean.getUserId() == null || ontologyId.isEmpty()
+					|| subscriptionsBean.getNotificationType() == null) {
+				throw new NotificationNotFoundException(
+						NotificationNotFoundException.DEFAULT_MESSAGE);
+			}
 			subscriptionsService.removeSubscriptions(subscriptionsBean);
+		} catch (NotificationNotFoundException notificationNotFoundException) {
+			response.setStatus(Status.CLIENT_ERROR_NOT_FOUND,
+					notificationNotFoundException.getMessage());
 		} catch (Exception e) {
 			response.setStatus(Status.SERVER_ERROR_INTERNAL, e.getMessage());
 			e.printStackTrace();
@@ -180,11 +208,21 @@ public class SubscritpionsRestlet extends AbstractBaseRestlet {
 	 * @param resp
 	 */
 	private void updateSubscriptions(Request request, Response response) {
-		
+
 		SubscriptionsBean subscriptionsBean = BeanHelper
 				.populateSubscriptionsBeanFromRequest(request);
+		// List For OntologyIds
+		List<String> ontologyId = subscriptionsBean.getOntologyIds();
 		try {
+			if (subscriptionsBean.getUserId() == null || ontologyId.isEmpty()
+					|| subscriptionsBean.getNotificationType() == null) {
+				throw new NotificationNotFoundException(
+						NotificationNotFoundException.DEFAULT_MESSAGE);
+			}
 			subscriptionsService.updateSubscriptions(subscriptionsBean);
+		} catch (NotificationNotFoundException notificationNotFoundException) {
+			response.setStatus(Status.CLIENT_ERROR_NOT_FOUND,
+					notificationNotFoundException.getMessage());
 		} catch (Exception e) {
 			response.setStatus(Status.SERVER_ERROR_INTERNAL, e.getMessage());
 			e.printStackTrace();
